Update the spatial model
We run the model until coalescence, then we wait N generations
(specified as a parameter), which we compute based on the mean age of
parents of individuals. Once the burn-in period has occured, we start
recording individuals on a specified interval of generations n times.
Finally, we write a tree sequences with the metadata about the recording
times and generation times over time.