This needed some cleanup
The aggregation of stats was not correct, we only want to sum across
replicates, with a masked diagonal. In addition, the aggregation of
stats after running was less than ideal, now we keep a running sum if we
plan to sum across replicates. This was validated against some code I
already had for running validation locally.